backend/requirements.txt
django
djangorestframework
mysqlclient
pillow
drf-yasg
